rename tools file
